requirements.txt
requests>=2.28.1, <3.0.0
browser_cookie3>=0.16.5, <1.0.0
tzdata
python-crontab>=2.7.1, <3.0.0